// projects/ng-dijta/src/lib/components/dx-side-bar/model/menu.ts
import { SUB_MENU_TYPE } from "../model";

export type MENU =
  | 'HOME'
  | 'PAYMENT'
  | 'QUOTE'
  | 'POLICY'
  | 'OPERATIONS'
  | 'REPORTS'
  | 'SETTINGS'
  | 'USER'
  | 'AGENTPROFILE'
  | 'AGENTDASHBOARD';
export type SUB_MENU =
  | 'PAYFORT'
  | 'COLLECTION'
  | 'QUOTE'
  | 'UNDERWRITING'
  | 'CANCELLATION'
  | 'ENQUIRY'
  | 'BRANCH'
  | 'AGENT'
  | 'CHANNEL'
  | 'REPORTS'
  | 'POLICYREPORTS'
  | 'ISSUANCEREPORTS'
  | 'PRODUCT'
  | 'PRODUCTCATEGORY'
  | 'GROUP'
  | 'ADDPAYMENT'
  | 'VERIFYPAYMENT'
  | 'PAYINSLIP'
  | 'PAYMENTMETHODS';

export interface Menu {
  label?: string;
  code?: string;
  active?: boolean;
  order?: number;
  route?: MenuRouter;
  icon?: string;
  position?: 'bottom';
  children?: SubMenu[];
  isToggleChildMenu?:boolean;
  resetMenu?:boolean;
}

export interface MenuRouter {
  path?: string;
  param?: { [key: string]: string | number };
}

export interface SubMenu extends Menu {
  parentId: string;
  parentName?: string;
}

// Original

export interface OriginalMenu {
  menu: string;
  submenu: Submenu[];
  menuLabel: string;
  menuIcon: string;
}

export interface Submenu {
  subMenuName: string;
  path: string;
  subMenuLabel: string;
  subMenuIcon: string;
  actions: string[];
}

export interface ActivePage {
  parentCode?: string;
  childCode?: string;
}

export interface Breadcrumb {
  title?: string;
  active?: boolean;
  routerLink?: string;
}

export interface ActiveMenu {
  menu: MENU;
  subMenu?: SUB_MENU;
}

export interface MenuModule {
  path?: any;
  menu?: string;
  menuIcon?: string;
  menuLabel?: string;
  position?: 'bottom';
  submenu?: Array<SubMenuModule>;
  order?: number
}
export interface SubMenuModule {
  actions?: Array<string>;
  path?: string;
  subMenuIcon?: string;
  subMenuLabel?: string;
  subMenuName?: string;
  order?: number
}

export interface FlexMenuAction {
  state?: SUB_MENU_TYPE;
  mouseOver?: boolean
}